requirements.txt
django=2.2
djangorestframework
psycopg2-binary
